requirements.txt

Flask-RESTful==0.3.9
Flask-Reuploaded==1.3.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1

keras==2.8.0
Keras-Preprocessing==1.1.2
libclang==14.0.6
Markdown==3.4.1
oauthlib==3.2.2
opencv-python==4.5.5.62

pywin32==305
requests-oauthlib==1.3.1
rsa==4.9
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-io-gcs-filesystem==0.29.0
termcolor==2.1.1
tf-estimator-nightly==2.8.0.dev2021122109
urllib3==1.26.13
WTForms==3.0.1
protobuf==3.20
